Move image based features to submodule
